# openapi.yaml
openapi: 3.0.3
info:
  title: Employee Management System API
  version: 1.0.0
  description: Employee Management System REST API (contract-first).
servers:
  - url: /api/v1
tags:
  - name: Auth
  - name: Employees
  - name: Departments
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/register:
    post:
      tags:
        - Auth
      summary: Register
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /employees:
    post:
      tags:
        - Employees
      summary: Create employee
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreateRequest'
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Employees
      summary: List employees
      operationId: listEmployees
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - name: departmentId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: role
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EmployeeRole'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EmployeeStatus'
      responses:
        '200':
          description: Employees listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageEmployee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /employees/{id}:
    parameters:
      - $ref: '#/components/parameters/EmployeeId'
    get:
      tags:
        - Employees
      summary: Get employee by ID
      operationId: getEmployeeById
      responses:
        '200':
          description: Employee found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Employees
      summary: Update employee
      operationId: updateEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdateRequest'
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Employees
      summary: Soft delete employee
      operationId: softDeleteEmployee
      responses:
        '204':
          description: Employee soft-deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /departments:
    post:
      tags:
        - Departments
      summary: Create department
      operationId: createDepartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentCreateRequest'
      responses:
        '201':
          description: Department created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Departments
      summary: List departments
      operationId: listDepartments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Departments listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageDepartment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /departments/{id}:
    parameters:
      - $ref: '#/components/parameters/DepartmentId'
    get:
      tags:
        - Departments
      summary: Get department by ID
      operationId: getDepartmentById
      responses:
        '200':
          description: Department found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Departments
      summary: Update department
      operationId: updateDepartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentUpdateRequest'
      responses:
        '200':
          description: Department updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Departments
      summary: Delete department
      operationId: deleteDepartment
      responses:
        '204':
          description: Department deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    Page:
      name: page
      in: query
      required: false
      description: 0-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Sort:
      name: sort
      in: query
      required: false
      description: Sort criteria in the format `field,(asc|desc)`
      schema:
        type: string
        example: createdAt,desc

    EmployeeId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

    DepartmentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    EmployeeRole:
      type: string
      enum:
        - ADMIN
        - HR
        - EMPLOYEE

    EmployeeStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE

    Employee:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - designation
        - salary
        - departmentId
        - role
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          minLength: 7
          maxLength: 20
          pattern: '^[0-9+()\-\s]{7,20}$'
          nullable: true
        designation:
          type: string
          minLength: 1
          maxLength: 100
        salary:
          type: number
          format: double
          minimum: 0
        departmentId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/EmployeeRole'
        status:
          $ref: '#/components/schemas/EmployeeStatus'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true

    EmployeeCreateRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - designation
        - salary
        - departmentId
        - role
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          minLength: 7
          maxLength: 20
          pattern: '^[0-9+()\-\s]{7,20}$'
          nullable: true
        designation:
          type: string
          minLength: 1
          maxLength: 100
        salary:
          type: number
          format: double
          minimum: 0
        departmentId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/EmployeeRole'
        status:
          $ref: '#/components/schemas/EmployeeStatus'

    EmployeeUpdateRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - designation
        - salary
        - departmentId
        - role
        - status
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          minLength: 7
          maxLength: 20
          pattern: '^[0-9+()\-\s]{7,20}$'
          nullable: true
        designation:
          type: string
          minLength: 1
          maxLength: 100
        salary:
          type: number
          format: double
          minimum: 0
        departmentId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/EmployeeRole'
        status:
          $ref: '#/components/schemas/EmployeeStatus'

    Department:
      type: object
      required:
        - id
        - name
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true

    DepartmentCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true

    DepartmentUpdateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
          nullable: true

    PageMeta:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0

    PageEmployee:
      type: object
      required:
        - items
        - meta
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        meta:
          $ref: '#/components/schemas/PageMeta'

    PageDepartment:
      type: object
      required:
        - items
        - meta
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Department'
        meta:
          $ref: '#/components/schemas/PageMeta'

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password

    RegisterRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
        - designation
        - salary
        - departmentId
        - role
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
        phone:
          type: string
          minLength: 7
          maxLength: 20
          pattern: '^[0-9+()\-\s]{7,20}$'
          nullable: true
        designation:
          type: string
          minLength: 1
          maxLength: 100
        salary:
          type: number
          format: double
          minimum: 0
        departmentId:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/EmployeeRole'

    AuthResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
      properties:
        accessToken:
          type: string
          minLength: 1
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          minimum: 1
          description: Token expiry in seconds

    ErrorDetail:
      type: object
      required:
        - field
        - issue
      properties:
        field:
          type: string
          minLength: 1
        issue:
          type: string
          minLength: 1

    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - code
        - message
        - path
        - correlationId
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          minimum: 100
          maximum: 599
        error:
          type: string
          minLength: 1
        code:
          type: string
          minLength: 1
          description: Stable, documented error identifier
          example: VALIDATION_ERROR
        message:
          type: string
          minLength: 1
        path:
          type: string
          minLength: 1
        correlationId:
          type: string
          minLength: 1
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'